separate APIs groups
